AMBARI-26000: Make Ambari supports Python3 environment (#3751)
* AMBARI-26000 support python3 on trunk
* remove yum since it's not supported in py3
* 1.fix shell decode error
2.fix sudo encoding error
3.remove redunant get_encoded_string
* remove duplicate import
* fix bugs
* update ambari metrics to python3
* fix review comment

+++ ambari-agent/src/main/python/ambari_agent/alerts/script_alert.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 """
 Licensed to the Apache Software Foundation (ASF) under one
@@ -55,7 +55,7 @@ class ScriptAlert(BaseAlert):
     self.parameters = {}
 
     # will force a kinit even if klist says there are valid tickets (4 hour default)
-    self.kinit_timeout = long(config.get('agent', 'alert_kinit_timeout', BaseAlert._DEFAULT_KINIT_TIMEOUT))
+    self.kinit_timeout = int(config.get('agent', 'alert_kinit_timeout', BaseAlert._DEFAULT_KINIT_TIMEOUT))
 
     if 'path' in alert_source_meta:
       self.path = alert_source_meta['path']
